Rename repo-url endpoint
